refactor: community 타입 파일 구조 정리
