+++ web/src/components/communityFeed/CommunityPostPreview.tsx
@@ -1,6 +1,6 @@
 import { useNavigate } from "react-router-dom";
 
-import type { CommunityPost } from "@/types/communityDetail/communityDetail.types";
+import type { CommunityPost } from "@/types/community/community.types";
 import { cn } from "@/utils/cn";
 
 import { NameTag } from "@/components/common/NameTag";
