+++ web/src/types/community/community.types.ts
@@ -20,3 +20,15 @@ export type Comment = {
   createdAt: string;
   children: Comment[];
 };
+
+export interface UploadedImage {
+  previewUrl: string;
+  file: File;
+}
+
+export type CommunityWriteValidationError = {
+  titleTooShort: boolean;
+  contentTooShort: boolean;
+  boardEmpty: boolean;
+  imageTooLarge: boolean;
+};
