--- web/src/types/communityWrite/communityWrite.types.ts
@@ -1,11 +1,0 @@
-export interface UploadedImage {
-  previewUrl: string;
-  file: File;
-}
-
-export type CommunityWriteValidationError = {
-  titleTooShort: boolean;
-  contentTooShort: boolean;
-  boardEmpty: boolean;
-  imageTooLarge: boolean;
-};
